+++ src/style/themes/mint/index.d.ts
@@ -1,5 +1,0 @@
-import { ThemeObject } from "../base";
-
-declare const mintTheme: ThemeObject;
-
-export default mintTheme;