+++ src/style/themes/none/none-theme.config.ts
@@ -1,0 +1,7 @@
+import { ThemeObject } from "../base";
+
+export default (): Partial<ThemeObject> => {
+  return {
+    name: "none",
+  };
+};